requirements.txt
crewai==0.51.0
crewai-tools==0.8.0
streamlit==1.38.0
python-dotenv==1.0.1
langchain-groq==0.2.0